[#6] test: split tests into unit/ and e2e/ with per-path scripts
