+++ .github/workflows/_checks.yml
@@ -14,5 +14,4 @@ jobs:
       - run: bun install --frozen-lockfile
       - run: bun run typecheck
       - run: bunx biome ci .
-      # e2e is gated behind RUN_E2E (needs ChatGPT OAuth + real generations), so this runs unit tests only.
-      - run: bun test
+      - run: bun run test
