+++ README.md
@@ -36,7 +36,7 @@ OpenCode auto-installs the package via Bun on next launch — no separate `npm i
 
 Just ask your agent in natural language and `gpt_imagegen` will be picked up.
 
-The three examples below are the **actual outputs of this repo's e2e test suite** — see [`tests/e2e.test.ts`](./tests/e2e.test.ts) for the exact prompts and assertions.
+The three examples below are the **actual outputs of this repo's e2e test suite** — see [`tests/e2e/subscription.test.ts`](./tests/e2e/subscription.test.ts) for the exact prompts and assertions.
 
 ### Example A — generate
