+++ tests/e2e/subscription.test.ts
@@ -8,7 +8,7 @@ import { pathToFileURL } from "node:url"
 
 const WORKDIR = await mkdtemp(path.join(os.tmpdir(), "qa-imagegen-work-"))
 const XDG_CONFIG_HOME = await mkdtemp(path.join(os.tmpdir(), "qa-imagegen-cfg-"))
-const REPO_DIR = path.resolve(import.meta.dir, "..")
+const REPO_DIR = path.resolve(import.meta.dir, "../..")
 const RUN_TIMEOUT_MS = 600_000
 const TEST_TIMEOUT_MS = RUN_TIMEOUT_MS + 10_000
 const STYLE = "hand-drawn 90s Japanese animation style"
@@ -78,7 +78,7 @@ function readPngDimensions(buf: Buffer): { width: number; height: number } {
   return { width: buf.readUInt32BE(16), height: buf.readUInt32BE(20) }
 }
 
-describe.skipIf(!process.env.RUN_E2E)("gpt_imagegen e2e", () => {
+describe("gpt_imagegen e2e (subscription)", () => {
   beforeAll(async () => {
     console.log(`WORKDIR: ${WORKDIR}`)
     console.log(`XDG_CONFIG_HOME: ${XDG_CONFIG_HOME}`)
