+++ tests/unit/output-image.test.ts
@@ -3,7 +3,7 @@ import { existsSync } from "node:fs"
 import { mkdtemp, readFile, rm, writeFile } from "node:fs/promises"
 import os from "node:os"
 import path from "node:path"
-import { buildSavedMessage, pickNonOverwritePath, saveGeneratedImage } from "../src/output-image"
+import { buildSavedMessage, pickNonOverwritePath, saveGeneratedImage } from "../../src/output-image"
 
 let dir: string
